examples/projects/option-greeks/requirements.txt
databento
pandas
scipy
pathway
python-dotenv
fastapi
pydantic
streamlit
uvicorn
